src/pilot/tools/system/screenshot_tool.py
"""
Cross-platform screenshot capture tool.
"""

import base64
import io
import time
from typing import Optional, Tuple, Dict, Any
from PIL import Image
import pyautogui
import platform


class ScreenshotTool:
    """
    Cross-platform screenshot capture with region support.
    Handles Retina/HiDPI display scaling automatically.
    Includes short-lived caching to avoid redundant captures.
    """

    CACHE_TTL = 0.0

    def __init__(self):
        """Initialize and detect display scaling."""
        self.scaling_factor = self._detect_scaling()
        self.os_type = platform.system().lower()
        self.active_window_bounds = None
        self._cache: Optional[Tuple[float, Optional[Tuple], Image.Image]] = None

    def _detect_scaling(self) -> float:
        """Detect display scaling factor (Retina = 2.0, normal = 1.0)."""
        screen_size = pyautogui.size()
        test_screenshot = pyautogui.screenshot()

        if test_screenshot.width > screen_size.width:
            scaling = test_screenshot.width / screen_size.width
            return scaling
        return 1.0

    def capture_active_window(
        self, app_name: Optional[str] = None
    ) -> Tuple[Image.Image, Dict[str, Any]]:
        """
        Capture ONLY the active window, not entire screen.

        Args:
            app_name: Optional app name to ensure correct window

        Returns:
            (PIL Image of window, dict with bounds and metadata)
        """
        try:
            if self.os_type == "darwin":
                return self._capture_active_window_macos(app_name)
            elif self.os_type == "windows":
                return self._capture_active_window_windows(app_name)
            else:
                return self._capture_active_window_linux(app_name)
        except Exception as e:
            raise RuntimeError(
                f"Failed to capture window for '{app_name}': {e}. "
                "Make sure the application is running and visible."
            )

    def _capture_active_window_macos(
        self, app_name: Optional[str] = None
    ) -> Tuple[Image.Image, Dict[str, Any]]:
        """Capture active window on macOS using Quartz."""
        from Quartz import (
            CGWindowListCopyWindowInfo,
            CGWindowListCreateImage,
            CGRectNull,
            kCGWindowListOptionIncludingWindow,
            kCGWindowListOptionOnScreenOnly,
            kCGNullWindowID,
            kCGWindowImageDefault,
        )
        import Quartz.CoreGraphics as CG

        window_list = CGWindowListCopyWindowInfo(
            kCGWindowListOptionOnScreenOnly, kCGNullWindowID
        )

        target_window = None
        for window in window_list:
            owner_name = window.get("kCGWindowOwnerName", "")
            layer = window.get("kCGWindowLayer", 999)

            if layer == 0:
                if app_name:
                    if app_name.lower() in owner_name.lower():
                        target_window = window
                        break
                else:
                    target_window = window
                    break

        if not target_window:
            available_apps = [
                w.get("kCGWindowOwnerName", "")
                for w in window_list
                if w.get("kCGWindowLayer", 999) == 0
            ]
            raise RuntimeError(
                f"Window '{app_name}' not found. Available: {available_apps[:5]}. "
                "Verify app is running and visible."
            )

        bounds = target_window["kCGWindowBounds"]
        x = int(bounds["X"])
        y = int(bounds["Y"])
        width = int(bounds["Width"])
        height = int(bounds["Height"])
        window_id = target_window["kCGWindowNumber"]

        cgimage = CGWindowListCreateImage(
            CGRectNull,
            kCGWindowListOptionIncludingWindow,
            window_id,
            kCGWindowImageDefault,
        )

        if not cgimage:
            region = (x, y, width, height)
            screenshot = self.capture(region=region)
        else:
            width_px = CG.CGImageGetWidth(cgimage)
            height_px = CG.CGImageGetHeight(cgimage)
            bytes_per_row = CG.CGImageGetBytesPerRow(cgimage)

            provider = CG.CGImageGetDataProvider(cgimage)
            data = CG.CGDataProviderCopyData(provider)

            import numpy as np

            byte_data = bytes(data)
            img_array = np.frombuffer(byte_data, dtype=np.uint8)
            img_array = img_array.reshape((height_px, bytes_per_row // 4, 4))
            img_array = img_array[:, :width_px, [2, 1, 0, 3]]
            screenshot = Image.fromarray(img_array, "RGBA")

        self.active_window_bounds = {
            "x": x,
            "y": y,
            "width": width,
            "height": height,
            "captured": True,
        }

        return screenshot, self.active_window_bounds

    def _capture_active_window_windows(
        self, app_name: Optional[str] = None
    ) -> Tuple[Image.Image, Dict[str, Any]]:
        """Capture active window on Windows."""
        return self.capture(), {
            "x": 0,
            "y": 0,
            "width": 0,
            "height": 0,
            "type": "fullscreen",
        }

    def _capture_active_window_linux(
        self, app_name: Optional[str] = None
    ) -> Tuple[Image.Image, Dict[str, Any]]:
        """Capture active window on Linux."""
        return self.capture(), {
            "x": 0,
            "y": 0,
            "width": 0,
            "height": 0,
            "type": "fullscreen",
        }

    def window_to_screen_coords(self, x: int, y: int) -> Tuple[int, int]:
        """
        Convert window-relative coordinates to screen-absolute coordinates.

        Args:
            x, y: Coordinates relative to window

        Returns:
            (x, y) in screen coordinates
        """
        if self.active_window_bounds:
            screen_x = self.active_window_bounds["x"] + x
            screen_y = self.active_window_bounds["y"] + y
            return (screen_x, screen_y)
        return (x, y)

    def capture(
        self,
        region: Optional[Tuple[int, int, int, int]] = None,
        use_cache: bool = True,
    ) -> Image.Image:
        """
        Capture screenshot of entire screen or specific region.

        Args:
            region: Optional region as (x, y, width, height) in SCREEN coordinates
            use_cache: Whether to use cached screenshot if available (default True)

        Returns:
            PIL Image object at full resolution
        """
        now = time.time()
        if use_cache and self._cache is not None:
            cache_time, cache_region, cache_image = self._cache
            if (now - cache_time) < self.CACHE_TTL and cache_region == region:
                return cache_image

        if region:
            x, y, w, h = region
            scaled_region = (
                int(x * self.scaling_factor),
                int(y * self.scaling_factor),
                int(w * self.scaling_factor),
                int(h * self.scaling_factor),
            )
            screenshot = pyautogui.screenshot(region=scaled_region)
        else:
            screenshot = pyautogui.screenshot()

        self._cache = (now, region, screenshot)
        return screenshot

    def invalidate_cache(self) -> None:
        """Clear the screenshot cache."""
        self._cache = None

    def capture_as_base64(
        self, region: Optional[Tuple[int, int, int, int]] = None, format: str = "PNG"
    ) -> str:
        """
        Capture screenshot and return as base64 string.

        Args:
            region: Optional region as (x, y, width, height)
            format: Image format (PNG, JPEG, etc.)

        Returns:
            Base64 encoded screenshot
        """
        screenshot = self.capture(region=region)

        buffer = io.BytesIO()
        screenshot.save(buffer, format=format)
        buffer.seek(0)

        encoded = base64.b64encode(buffer.read()).decode("utf-8")
        return encoded

    def save(
        self, filepath: str, region: Optional[Tuple[int, int, int, int]] = None
    ) -> str:
        """
        Capture and save screenshot to file.

        Args:
            filepath: Path to save screenshot
            region: Optional region as (x, y, width, height)

        Returns:
            Path where screenshot was saved
        """
        screenshot = self.capture(region=region)
        screenshot.save(filepath)
        return filepath

    def get_pixel_color(self, x: int, y: int) -> Tuple[int, int, int]:
        """
        Get RGB color of pixel at specified coordinates.

        Args:
            x: X coordinate
            y: Y coordinate

        Returns:
            RGB tuple (r, g, b)
        """
        screenshot = self.capture()
        return screenshot.getpixel((x, y))
