pyproject.toml
[project]
name = "pilot"
version = "0.1.0"
description = "Cross-platform computer use agent with multi-tier accuracy system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai[tools]>=1.6.0",
    "browser-use>=0.9.4",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.8.2",
    "psutil>=5.9.0",
    "python-dotenv>=1.0.0",
    "opencv-python>=4.8.0",
    "paddlepaddle>=2.6.0",
    "paddleocr>=2.7.0",
    "easyocr>=1.7.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.0.10",
    "rich>=14.2.0",
    "prompt-toolkit>=3.0.43",
    "twilio>=9.0.0",
    "flask>=3.0.0",
    "pyngrok>=7.0.0",
    "pynput>=1.7.6",
    "deepgram-sdk>=5.3.0",
    "sounddevice>=0.4.6",
    "google-genai>=1.0.0",
    "inquirerpy>=0.3.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
macos = [
    "atomacos>=0.5.0",
    "pyobjc-framework-ApplicationServices>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-Vision>=10.0",
]
windows = [
    "pywinauto>=0.6.8",
    "comtypes>=1.2.0",
]
linux = [
    "python3-pyatspi",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pilot"]

[dependency-groups]
dev = [
    "hypothesis>=6.150.3",
    "pyte>=0.8.2",
    "pytest>=9.0.1",
]

